% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{calc_incidence_individual}
\alias{calc_incidence_individual}
\title{Calculate Incidence for Individual Pathogen (with DOW adjustment)}
\usage{
calc_incidence_individual(
  a,
  time_idx,
  pathogen_idx,
  post,
  components,
  dow_effect
)
}
\arguments{
\item{a}{Array of log-incidence posterior samples [samples, pathogens, time]}

\item{time_idx}{Integer time index}

\item{pathogen_idx}{Integer pathogen index}

\item{post}{Posterior samples object containing day_of_week_simplex if DOW effects used}

\item{components}{Model components containing DOW information}

\item{dow_effect}{Logical whether day-of-week effect to be calculated}
}
\value{
Vector of incidence posterior samples
}
\description{
Computes incidence for a specific pathogen, applying day-of-week effects if present
}
