% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{calc_incidence_single}
\alias{calc_incidence_single}
\title{Calculate Incidence for Single Pathogen Model (with DOW adjustment)}
\usage{
calc_incidence_single(a, time_idx, pathogen_idx, post, components)
}
\arguments{
\item{a}{Array of log-incidence posterior samples [samples, time]}

\item{time_idx}{Integer time index}

\item{pathogen_idx}{NULL (unused but required for interface consistency)}

\item{post}{Posterior samples object containing day_of_week_simplex if DOW effects used}

\item{components}{Model components containing DOW information}
}
\value{
Vector of incidence posterior samples
}
\description{
Computes incidence for single pathogen models, applying day-of-week effects if present
}
