% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{calc_incidence_total}
\alias{calc_incidence_total}
\title{Calculate Total Incidence Across All Pathogens (with DOW adjustment)}
\usage{
calc_incidence_total(a, time_idx, pathogen_idx, post, components, dow_effect)
}
\arguments{
\item{a}{Array of log-incidence posterior samples [samples, pathogens, time]}

\item{time_idx}{Integer time index}

\item{pathogen_idx}{NULL (unused but required for interface consistency)}

\item{post}{Posterior samples object containing day_of_week_simplex if DOW effects used}

\item{components}{Model components containing DOW information}

\item{dow_effect}{Logical whether day-of-week effect to be calculated}
}
\value{
Vector of total incidence posterior samples
}
\description{
Computes total incidence across all pathogens, applying day-of-week effects if present
}
