% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion.R
\name{calc_proportion}
\alias{calc_proportion}
\title{Calculate proportion for Individual Pathogen}
\usage{
calc_proportion(
  a,
  time_idx,
  pathogen_idx,
  post,
  components,
  numerator_idx,
  denominator_idx
)
}
\arguments{
\item{a}{Array of log-incidence posterior samples [samples, pathogens, time]}

\item{time_idx}{Integer time index}

\item{pathogen_idx}{Integer pathogen index}

\item{post}{Posterior samples object (unused but required for interface consistency)}

\item{components}{Model components (unused but required for interface consistency)}

\item{numerator_idx}{Integer index of the pathogen/s used in numerator of proportion}

\item{denominator_idx}{Integer index of the pathogen/s used in denominator of proportion}
}
\value{
Vector of proportion posterior samples
}
\description{
Computes log-difference in incidence for a specific pathogen
}
