% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiStrainDynamics-package.R
\name{EpiStrainDynamics-package}
\alias{EpiStrainDynamics-package}
\alias{EpiStrainDynamics}
\title{The 'EpiStrainDynamics' package.}
\description{
A statistical framework developed for inferring temporal trends of multiple pathogens from routinely collected surveillance data.
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.3. https://mc-stan.org
}
