% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_fit_calc.R
\name{expand_pathogen_grid}
\alias{expand_pathogen_grid}
\title{Expand Time Grid for Multiple Pathogens}
\usage{
expand_pathogen_grid(time_grid, pathogen_names)
}
\arguments{
\item{time_grid}{Data frame from with time_idx columns}

\item{pathogen_names}{Character vector of pathogen names}
}
\value{
Data frame with columns: pathogen, pathogen_idx, time_idx
}
\description{
Expands a time grid to include all pathogen-time combinations (sequential
pairing, not Cartesian product)
}
