% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{extract_smoothing_parameters}
\alias{extract_smoothing_parameters}
\title{Extract smoothing parameters from P-spline models}
\usage{
extract_smoothing_parameters(fitted_model)
}
\arguments{
\item{fitted_model}{fitted P-spline model object}
}
\value{
list of smoothing parameter estimates
}
\description{
Extract smoothing parameters from P-spline models
}
