% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit}
\alias{fit}
\title{Method fit}
\usage{
fit(constructed_model, iter, warmup, chains)
}
\arguments{
\item{constructed_model}{prepared model object from construct_model()}

\item{iter}{number of iterations}

\item{warmup}{number of warmup iterations}

\item{chains}{number of chains}
}
\value{
fit model
}
\description{
Method fit
}
