% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_model.R
\name{get_knots}
\alias{get_knots}
\title{Function for getting knot locations}
\usage{
get_knots(X, days_per_knot = 3, spline_degree = 3)
}
\arguments{
\item{X}{Numeric vector of time points}

\item{days_per_knot}{Number of days between knots (must be positive)}

\item{spline_degree}{Polynomial degree of spline (must be positive)}
}
\value{
Numeric vector of knot locations
}
\description{
Function for getting knot locations
}
