% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_model.R
\name{get_model_type}
\alias{get_model_type}
\title{Extract model type from method and pathogen structure}
\usage{
get_model_type(method_name, pathogen_type)
}
\arguments{
\item{method_name}{Character string: method name ('random-walk' or 'p-spline')}

\item{pathogen_type}{Character string: pathogen structure type}
}
\value{
Character string: model type
}
\description{
Extract model type from method and pathogen structure
}
