% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogen_structure.R
\name{get_noise_structure}
\alias{get_noise_structure}
\title{Get noise structure from specified observation noise}
\usage{
get_noise_structure(
  observation_noise = c("observation_noise_only", "pathogen_specific_noise")
)
}
\arguments{
\item{observation_noise}{either 'observation_noise_only' (only includes
observation noise - the same between pathogens) or pathogen_specific_noise'
(includes noise in individual pathogens as well)}
}
\value{
numeric value for noise structure needed by stan models
}
\description{
Get noise structure from specified observation noise
}
