% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_rate.R
\name{gr_internal}
\alias{gr_internal}
\title{Internal growth rate function}
\usage{
gr_internal(iter, obj, time_labels)
}
\arguments{
\item{iter}{loop id}

\item{obj}{posterior samples}

\item{time_labels}{time values}
}
\value{
row processed for given iter
}
\description{
Internal growth rate function
}
