% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{incidence.ps}
\alias{incidence.ps}
\title{Incidence Analysis for Penalized Spline Multi-Pathogen Models}
\usage{
\method{incidence}{ps}(fitted_model, ...)
}
\arguments{
\item{fitted_model}{Fitted model object of class 'ps'}

\item{...}{Additional arguments (unused)}
}
\value{
Data frame with incidence results for individual pathogens and total
}
\description{
Incidence Analysis for Penalized Spline Multi-Pathogen Models
}
