% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{incidence.ps_single}
\alias{incidence.ps_single}
\title{Incidence Analysis for Penalized Spline Single Pathogen Models}
\usage{
\method{incidence}{ps_single}(fitted_model, ...)
}
\arguments{
\item{fitted_model}{Fitted model object of class 'ps_single'}

\item{...}{Additional arguments (unused)}
}
\value{
Data frame with incidence results
}
\description{
Incidence Analysis for Penalized Spline Single Pathogen Models
}
