% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{incidence.rw}
\alias{incidence.rw}
\title{Incidence Analysis for Random Walk Multi-Pathogen Models}
\usage{
\method{incidence}{rw}(fitted_model, ...)
}
\arguments{
\item{fitted_model}{Fitted model object of class 'rw'}

\item{...}{Additional arguments (unused)}
}
\value{
Data frame with incidence results for individual pathogens and total
}
\description{
Incidence Analysis for Random Walk Multi-Pathogen Models
}
