% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{incidence.rw_single}
\alias{incidence.rw_single}
\title{Incidence Analysis for Random Walk Single Pathogen Models}
\usage{
\method{incidence}{rw_single}(fitted_model, ...)
}
\arguments{
\item{fitted_model}{Fitted model object of class 'rw_single'}

\item{...}{Additional arguments (unused)}
}
\value{
Data frame with incidence results
}
\description{
Incidence Analysis for Random Walk Single Pathogen Models
}
