% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{influenza}
\alias{influenza}
\title{THIS NEEDS TO BE EDITED}
\format{
## `influenza`
A data frame with 830 rows and 15 columns:
\describe{
  \item{week}{Date, between 2 January 2012 to week starting 25 December 2023}
  \item{ili}{Integer, daily total number of cases of influenza-like illness}
  \item{inf_A}{Integer, daily number of cases of unsubtyped influenza A}
  \item{inf_B}{Integer, daily number of cases of influenza B}
  \item{inf_all}{Integer, daily number of cases of }
  \item{inf_neg}{Integer, daily number of cases of }
  \item{inf_H3N2}{Integer, daily number of cases of influenza A subtype H3N2}
  \item{inf_H1N1}{Integer, daily number of cases of influenza A subtype H1N1}
  \item{num_spec}{Integer, number of specimens} ??
}
}
\source{
<https://www.who.int/teams/global-influenza-programme/surveillance-and-monitoring/influenza-surveillance-outputs>
}
\usage{
influenza
}
\description{
Influenza-like illness data were retrieved from the World Health
Organization’s Global Influenza Programme.16 We collated weekly data for
Australia from the week starting January 2, 2012 to the week starting
December 25, 2023 inclusive. The data described: (1) the weekly number
of cases of influenza-like illness; and (2) the weekly number of specimens
positive for influenza by subtype (and the number of negative tests). We
grouped the influenza specimens into: influenza A subtype not determined;
influenza A H3N2; influenza A H1N1 (influenza A H1N1, influenza A H1N1pdm09);
influenza B (influenza B Yamagata, influenza B Victoria, influenza B lineage
not determined); and negative tests. Negative tests likely reflect
individuals infected with other ILI-causing pathogens or other unknown
causes which we do not have data to determine.
}
\keyword{datasets}
