% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogen_structure.R
\name{multiple}
\alias{multiple}
\title{Multiple pathogen structure}
\usage{
multiple(case_timeseries, time, component_pathogen_timeseries)
}
\arguments{
\item{case_timeseries}{vector of total case data}

\item{time}{vector of labelled time data}

\item{component_pathogen_timeseries}{named list of each pathogen that has
timeseries of case counts}
}
\value{
named list including pathogen_structure, pathogen_names, and data
of class \code{EpiStrainDynamics.pathogen_structure}
}
\description{
Multiple pathogen structure
}
\examples{
multiple(
  case_timeseries = sarscov2$cases,
  time = sarscov2$date,
  component_pathogen_timeseries = list(
    alpha = sarscov2$alpha,
    delta = sarscov2$delta,
    omicron = sarscov2$omicron,
    other = sarscov2$other
  )
)

}
\seealso{
Other pathogen_structure: 
\code{\link{single}()},
\code{\link{subtyped}()}
}
\concept{pathogen_structure}
