% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.incidence}
\alias{plot.growth_rate}
\alias{plot.Rt}
\alias{plot.proportion}
\title{Plot metrics calculation outputs}
\usage{
\method{plot}{incidence}(x, xlab = "Time", ...)

\method{plot}{growth_rate}(x, xlab = "Time", ...)

\method{plot}{Rt}(x, xlab = "Time", ...)

\method{plot}{proportion}(x, xlab = "Time", ...)
}
\arguments{
\item{x}{Metrics calculation output of class \code{EpiStrainDynamics.metric}}

\item{xlab}{Time label for x axis, defaults to "Time"}

\item{...}{Additional arguments passed to plot}
}
\value{
ggplot2 plot output
}
\description{
S3 methods for plotting metrics calculated from the output sof either
\code{\link[=incidence]{incidence()}}, \code{\link[=growth_rate]{growth_rate()}}, \code{\link[=Rt]{Rt()}}, or \code{\link[=proportion]{proportion()}}.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
mod <- construct_model(
  pathogen_structure = single(
    case_timeseries = sarscov2$cases,
    time = sarscov2$date
  ),
  method = random_walk()
)

fit <- fit_model(mod)
gr <- growth_rate(mod)
plot(gr)
\dontshow{\}) # examplesIf}
}
