% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_fit_calc.R
\name{predict_B_true}
\alias{predict_B_true}
\title{Predict B-spline Basis Matrix}
\usage{
predict_B_true(time_seq, days_per_knot, spline_degree)
}
\arguments{
\item{time_seq}{Numeric vector of time points}

\item{days_per_knot}{Integer number of days between knots}

\item{spline_degree}{Integer degree of B-splines}
}
\value{
Matrix B_true for transforming spline coefficients
}
\description{
Creates B-spline basis matrix for transforming spline coefficients to time series
}
