% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_mp_stan.R
\name{ps_mp_stan}
\alias{ps_mp_stan}
\title{PS MP}
\usage{
ps_mp_stan(
  num_data,
  num_knots,
  num_path,
  knots,
  spline_degree,
  Y,
  P,
  X,
  week_effect,
  DOW,
  cov_structure = c(0, 1, 2),
  ...
)
}
\arguments{
\item{num_data}{number of data points}

\item{num_knots}{num of knots}

\item{knots}{the sequence of knots}

\item{spline_degree}{the degree of spline (is equal to order - 1)}

\item{week_effect}{Number of days in day of week effect? 1=none, 2=weekends?, 7=all days}

\item{cov_structure}{0 is tau\link{1}, 1 is tau\link{num_path}, 2 is Sigma\link{num_path, num_path}}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}

\item{DOW[num_data]}{integer of day of the week}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
PS MP
}
