% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_single_stan.R
\name{ps_single_stan}
\alias{ps_single_stan}
\title{PS single}
\usage{
ps_single_stan(
  num_data,
  num_knots,
  knots,
  spline_degree,
  Y,
  X,
  week_effect,
  DOW,
  ...
)
}
\arguments{
\item{num_data}{number of data points}

\item{num_knots}{number of knots}

\item{knots}{the sequence of knots}

\item{spline_degree}{the degree of spline (is equal to order - 1)}

\item{Y}{outcome variable, eg daily number of cases}

\item{X}{time data}

\item{week_effect}{Number of distinct days in day of week effect. 1 = single
effect shared by all days (essentially no DOW), 7 = each day a unique effect}

\item{DOW}{integer of day of the week}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
PS single
}
