% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{random_walk}
\alias{random_walk}
\title{Specify random walk method}
\usage{
random_walk()
}
\value{
list with method identified as random walk of class
  `EpiStrainDynamics.method`
}
\description{
Specify random walk method
}
\examples{
random_walk()
}
\seealso{
Other method: 
\code{\link{p_spline}()}
}
\concept{method}
