% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_mp_stan.R
\name{rw_mp_stan}
\alias{rw_mp_stan}
\title{RW MP}
\usage{
rw_mp_stan(
  num_data,
  num_path,
  Y,
  P,
  week_effect,
  DOW,
  cov_structure = c(0, 1, 2),
  noise_structure = c(0, 1),
  ...
)
}
\arguments{
\item{num_data}{number of data points}

\item{num_path}{number of pathogens}

\item{Y}{outcome variable, eg daily number of cases}

\item{P}{daily number of lab tests positive for each pathogen}

\item{week_effect}{Number of distinct days in day of week effect. 1 = single
effect shared by all days (essentially no DOW), 7 = each day a unique effect}

\item{DOW}{integer of day of the week}

\item{cov_structure}{0 is tau of 1, 1 is tau of
num_path, 2 is Sigma with two dimensions: num_path, num_path.}

\item{noise_structure}{0 only includes observation noise (same between
pathogens), 1 includes noise in individual pathogens as well}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
RW MP
}
