% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_single_stan.R
\name{rw_single_stan}
\alias{rw_single_stan}
\title{RW single}
\usage{
rw_single_stan(num_data, Y, week_effect, DOW, ...)
}
\arguments{
\item{num_data}{number of data points}

\item{Y}{outcome variable, eg daily number of cases}

\item{week_effect}{Number of distinct days in day of week effect. 1 = single
effect shared by all days (essentially no DOW), 7 = each day a unique effect}

\item{DOW}{integer of day of the week}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
RW single
}
