% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogen_structure.R
\name{single}
\alias{single}
\title{Single pathogen structure}
\usage{
single(data, case_timeseries, time)
}
\arguments{
\item{data}{dataframe containing columns with all relevant data}

\item{case_timeseries}{name of column with total case data, this name will be
used as the name of the pathogen}

\item{time}{name of column with time data. flexible format - can be date,
index, or others, accepted as \code{index} identifiers in the \code{tsibble} time
format.}
}
\value{
formatted list with pathogen structure and data of class
\code{EpiStrainDynamics.pathogen_structure}.
}
\description{
Single pathogen structure
}
\examples{
single(
  data = sarscov2,
  case_timeseries = 'cases',
  time = 'date'
)

}
\seealso{
Other pathogen_structure: 
\code{\link{multiple}()},
\code{\link{subtyped}()}
}
\concept{pathogen_structure}
