% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogen_structure.R
\name{subtyped}
\alias{subtyped}
\title{Subtyped pathogen structure}
\usage{
subtyped(
  data,
  case_timeseries,
  unsubtyped_timeseries,
  subtyped_timeseries,
  other_pathogen_timeseries,
  time = NULL
)
}
\arguments{
\item{data}{dataframe containing columns with all relevant data, or a time
series object (ts, xts, zoo, tsibble, etc.)}

\item{case_timeseries}{Column name containing case counts. Must be numeric
or a \code{units} object from the \pkg{units} package.}

\item{unsubtyped_timeseries}{vector of column names with additional
unsubtyped pathogen case count timeseries. Must be numeric or a
\code{units} object from the \pkg{units} package.}

\item{subtyped_timeseries}{vector of column names with additional
subtyped pathogen case count timeseries. Must be numeric or a
\code{units} object from the \pkg{units} package.}

\item{other_pathogen_timeseries}{vector of column names with additional
pathogen case count timeseries to model. Must be numeric or a \code{units}
object from the \pkg{units} package.}

\item{time}{name of column with time data. Required for non-time-series
input data. Flexible format - can be date, index, or others, accepted
as \code{index} identifiers in the \code{tsibble} time format. Optional when
\code{data} is a time series class object (ts, mts, xts, zoo, zooreg, tsibble)
as the time index will be automatically detected.}
}
\value{
named list including pathogen_structure, pathogen_names, and data
of class \code{EpiStrainDynamics.pathogen_structure}
}
\description{
Subtyped pathogen structure
}
\examples{
# Using a data frame
subtyped(
  data = influenza,
  case_timeseries = "ili",
  unsubtyped_timeseries = "inf_A",
  subtyped_timeseries = c("inf_H3N2", "inf_H1N1"),
  other_pathogen_timeseries = c("inf_B", "other"),
  time = "week"
)

\dontshow{if (rlang::is_installed("xts")) withAutoprint(\{ # examplesIf}
# Using a time series object (time argument is optional)
influenza_xts <- xts::xts(
  influenza[, c("ili", "inf_A", "inf_H3N2", "inf_H1N1", "inf_B", "other")],
  order.by = influenza$week
)
subtyped(
  data = influenza_xts,
  case_timeseries = "ili",
  unsubtyped_timeseries = "inf_A",
  subtyped_timeseries = c("inf_H3N2", "inf_H1N1"),
  other_pathogen_timeseries = c("inf_B", "other")
)
\dontshow{\}) # examplesIf}
}
\seealso{
Other pathogen_structure: 
\code{\link{multiple}()},
\code{\link{single}()}
}
\concept{pathogen_structure}
