% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_fit_calc.R
\name{transform_posterior_multi}
\alias{transform_posterior_multi}
\title{Transform Posterior Samples Using Splines (Multiple Pathogens)}
\usage{
transform_posterior_multi(post, B_true, num_path, num_days)
}
\arguments{
\item{post}{Posterior samples object containing spline coefficients}

\item{B_true}{B-spline basis matrix from predict_B_true()}

\item{num_path}{Integer number of pathogens}

\item{num_days}{Integer number of time points}
}
\value{
Array of transformed posterior samples [samples, pathogens, time]
}
\description{
Transforms spline coefficients to time series for multiple pathogen models
}
