% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_fit_calc.R
\name{transform_posterior_single}
\alias{transform_posterior_single}
\title{Transform Posterior Samples Using Splines (Single Pathogen)}
\usage{
transform_posterior_single(post, B_true, num_days)
}
\arguments{
\item{post}{Posterior samples object containing spline coefficients}

\item{B_true}{B-spline basis matrix from predict_B_true()}

\item{num_days}{Integer number of time points}
}
\value{
Array of transformed posterior samples [samples, time]
}
\description{
Transforms spline coefficients to time series for single pathogen models
}
