% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-utils.R
\name{validate_list_vector_lengths}
\alias{validate_list_vector_lengths}
\title{Validate that a list of vectors all have the same length}
\usage{
validate_list_vector_lengths(
  vector_list,
  list_name,
  reference_length = NULL,
  .error_call = rlang::caller_env()
)
}
\arguments{
\item{vector_list}{Named list of vectors}

\item{list_name}{Name of the list parameter for error messages}

\item{reference_length}{Expected length (optional)}

\item{.error_call}{The calling function for better error messages}
}
\description{
Validate that a list of vectors all have the same length
}
