% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-utils.R
\name{validate_matching_lengths}
\alias{validate_matching_lengths}
\title{Validate input vectors have matching lengths}
\usage{
validate_matching_lengths(..., .error_call = rlang::caller_env())
}
\arguments{
\item{...}{Named vectors to check for matching lengths}

\item{.error_call}{The calling function for better error messages}
}
\description{
Validate input vectors have matching lengths
}
