% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_model.R
\name{construct_model}
\alias{construct_model}
\title{Construct model}
\usage{
construct_model(method, pathogen_structure, dow_effect = FALSE)
}
\arguments{
\item{method}{either random_walk() or p_spline()}

\item{pathogen_structure}{either single(), multiple(), or subtyped()}

\item{dow_effect}{logical whether to incorporate a day of week model}
}
\value{
a list containing the data, the model parameters, and pathogen
 names
}
\description{
Construct model
}
\examples{
mod <- construct_model(
  method = random_walk(),
  pathogen_structure = single(
    case_timeseries = sarscov2$cases,
    time = sarscov2$date,
    pathogen_name = 'SARS-COV-2'
  ),
  dow_effect = TRUE
)
}
