% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_model.R
\name{diagnose_model}
\alias{diagnose_model}
\title{Diagnose model convergence and fit}
\usage{
diagnose_model(fitted_model, rhat_threshold = 1.1, eff_sample_threshold = 100)
}
\arguments{
\item{fitted_model}{fitted model object}

\item{rhat_threshold}{R-hat threshold for convergence (default 1.1)}

\item{eff_sample_threshold}{effective sample size threshold}
}
\value{
list of diagnostic information
}
\description{
Diagnose model convergence and fit
}
\examples{
\dontrun{
  mod <- construct_model(
    method = random_walk(),
    pathogen_structure = single(
      case_timeseries = sarscov2$cases,
      time = sarscov2$date))
  fit <- fit_model(mod)
  diagnose(fit)
}

}
