% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_dispersion_structure.r
\name{dispersion_structure}
\alias{dispersion_structure}
\title{Create Dispersion Structure Specification}
\usage{
dispersion_structure(phi_mean = NULL, phi_sd = NULL)
}
\arguments{
\item{phi_mean}{Numeric scalar specifying the prior mean for the negative
binomial dispersion parameter. Must be positive. Optional - if not
provided, no priors will be set.}

\item{phi_sd}{Numeric scalar specifying the prior standard deviation for the
dispersion parameter. Must be positive. Optional - if not provided, no
priors will be set.}
}
\value{
An object of class \code{EpiStrainDynamics.dispersion} containing prior
specifications for phi parameter
}
\description{
This function creates a dispersion structure object that specifies priors
for the overdispersion parameter of the negative binomial likelihood for
the case timeseries.
}
\examples{
# Create dispersion structure
disp_struct <- dispersion_structure(phi_mean = 2.0, phi_sd = 0.5)

}
