% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit model}
\usage{
fit_model(constructed_model, iter = 2000, warmup = 1000, chains = 3)
}
\arguments{
\item{constructed_model}{prepared model object from construct_model()}

\item{iter}{number of iterations}

\item{warmup}{number of warmup iterations}

\item{chains}{number of chains}
}
\value{
fit model
}
\description{
Fit model
}
