% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_knots.R
\name{get_knots}
\alias{get_knots}
\title{Function for getting knot locations}
\usage{
get_knots(X, days_per_knot, spline_degree = 3)
}
\arguments{
\item{X}{days}

\item{days_per_knot}{days per knot}

\item{spline_degree}{spline degree}
}
\value{
location of knots
}
\description{
Function for getting knot locations
}
