% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_model.R
\name{get_model_type}
\alias{get_model_type}
\title{Extract model type from method and pathogen structure}
\usage{
get_model_type(method, pathogens)
}
\arguments{
\item{method}{method function, either random_walk() or p_spline()}

\item{pathogens}{pathogen structure, either single, multiple, or subtyped}
}
\value{
model type
}
\description{
Extract model type from method and pathogen structure
}
