% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_rate.R
\name{growth_rate}
\alias{growth_rate}
\alias{growth_rate.ps}
\alias{growth_rate.rw}
\alias{growth_rate.ps_single}
\alias{growth_rate.rw_single}
\title{Generic Method for Growth Rate Analysis}
\usage{
growth_rate(fitted_model, ...)

\method{growth_rate}{ps}(fitted_model, ...)

\method{growth_rate}{rw}(fitted_model, ...)

\method{growth_rate}{ps_single}(fitted_model, ...)

\method{growth_rate}{rw_single}(fitted_model, ...)
}
\arguments{
\item{fitted_model}{Fitted model object with appropriate class}

\item{...}{Additional arguments passed to methods}
}
\value{
Data frame with growth rate analysis results
}
\description{
S3 generic for computing growth rates from fitted models
}
