% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_rate.R
\name{growth_rate}
\alias{growth_rate}
\title{Calculate growth rate}
\usage{
growth_rate(fitted_model)
}
\arguments{
\item{fitted_model}{fitted model output created with \code{fit_model}}
}
\value{
growth rate calculation
}
\description{
Calculate growth rate
}
