% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{incidence}
\alias{incidence}
\alias{incidence.ps}
\alias{incidence.rw}
\alias{incidence.ps_single}
\alias{incidence.rw_single}
\title{Generic Method for Incidence Analysis}
\usage{
incidence(fitted_model, dow, ...)

\method{incidence}{ps}(fitted_model, dow, ...)

\method{incidence}{rw}(fitted_model, dow, ...)

\method{incidence}{ps_single}(fitted_model, dow, ...)

\method{incidence}{rw_single}(fitted_model, dow, ...)
}
\arguments{
\item{fitted_model}{Fitted model object with class `EpiStrainDynamics.fit`}

\item{dow}{Logical whether or not to include day-of-week in incidence calc}

\item{...}{Additional arguments passed to metrics calculation}
}
\value{
named list of class `EpiStrainDynamics.metric` containing a dataframe
 of the calculated metric outcome (`$measure`), the fit object (`$fit`), and the
 constructed model object (`$constructed_model`). The `measure` data frame
 contains the median of the epidemiological quantity (`y`), the 50% credible
 interval of the quantity (`lb_50` & `ub_50`), the 95% credible interval
 (`lb_95` & `ub_95`), the proportion greater than a defined threshold value
 (`prop`), the pathogen name (`pathogen`), and the time label (`time`).
}
\description{
S3 generic for computing incidence from fitted models
}
\examples{
\dontrun{
  mod <- construct_model(
    method = random_walk(),
    pathogen_structure = single(
      case_timeseries = sarscov2$cases,
      time = sarscov2$date))

  fit <- fit_model(mod)
  inc <- incidence(fit, dow = TRUE)
}
}
