% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogen_structure.R
\name{multiple}
\alias{multiple}
\title{Multiple pathogen structure}
\usage{
multiple(data, case_timeseries, time, component_pathogen_timeseries)
}
\arguments{
\item{data}{dataframe containing columns with all relevant data}

\item{case_timeseries}{name of column with total case data, this name will be
used as the name of the pathogen}

\item{time}{name of column with time data. flexible format - can be date,
index, or others, accepted as \code{index} identifiers in the \code{tsibble} time
format.}

\item{component_pathogen_timeseries}{vector of column names with additional
pathogen case count timeseries to model}
}
\value{
named list including pathogen_structure, pathogen_names, and data
of class \code{EpiStrainDynamics.pathogen_structure}
}
\description{
Multiple pathogen structure
}
\examples{
multiple(
  data = sarscov2,
  case_timeseries = 'cases',
  time = 'date',
  component_pathogen_timeseries = c('alpha', 'delta', 'omicron', 'other')
)

}
\seealso{
Other pathogen_structure: 
\code{\link{single}()},
\code{\link{subtyped}()}
}
\concept{pathogen_structure}
