% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.incidence}
\alias{plot.growth_rate}
\alias{plot.Rt}
\title{Generic Method for plotting}
\usage{
plot(df, ...)

\method{plot}{incidence}(df)

\method{plot}{growth_rate}(df)

\method{plot}{Rt}(df)
}
\arguments{
\item{df}{Fitted model object with appropriate class}

\item{...}{Additional arguments passed to methods}
}
\value{
Data frame with proportion analysis results
}
\description{
S3 generic for plotting
}
