% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.incidence}
\alias{plot.growth_rate}
\alias{plot.Rt}
\alias{plot.proportion}
\title{Generic Method for plotting metrics calculation outputs}
\usage{
plot(df, xlab = "Time", ...)

\method{plot}{incidence}(df, xlab = "Time", ...)

\method{plot}{growth_rate}(df, xlab = "Time", ...)

\method{plot}{Rt}(df, xlab = "Time", ...)

\method{plot}{proportion}(df, xlab = "Time", ...)
}
\arguments{
\item{df}{Metrics calculation output of class \code{EpiStrainDynamics.metric}
from either \code{incidence()}, \code{growth_rate()}, \code{Rt()}, or \code{proportion()}.}

\item{xlab}{Time label for x axis, defaults to "Time"}

\item{...}{Additional arguments passed to plot}
}
\value{
ggplot2 plot output
}
\description{
S3 generic for plotting
}
\examples{
\dontrun{
  mod <- construct_model(
    method = random_walk(),
    pathogen_structure = single(
      case_timeseries = sarscov2$cases,
      time = sarscov2$date))

  fit <- fit_model(mod)
  gr <- growth_rate(mod)
  plot(gr)
}
}
