% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion.R
\name{proportion}
\alias{proportion}
\title{Get proportion from model fit}
\usage{
proportion(
  fit,
  X = NULL,
  method = c("rw", "ps"),
  num_days = length(X),
  time_labels,
  days_per_knot = 5,
  spline_degree = 3,
  num_path = 4,
  comb_num = list(c(1, 2, 3), c(1, 2), c(3), c(1), c(2)),
  comb_den = list(c(1, 2, 3, 4), c(1, 2, 3, 4), c(1, 2, 3, 4), c(1, 2), c(1, 2)),
  comb_names = c("Influenza", "Influenza A", "Influenza B", "H3N2", "H1N1")
)
}
\arguments{
\item{fit}{model fit}

\item{method}{ps or rw}

\item{comb_names}{}
}
\description{
Get proportion from model fit
}
