% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion.R
\name{proportion}
\alias{proportion}
\alias{proportion.ps}
\alias{proportion.rw}
\alias{proportion.ps_single}
\alias{proportion.rw_single}
\title{Generic Method for proportion Analysis}
\usage{
proportion(
  fitted_model,
  numerator_combination = NULL,
  denominator_combination = NULL,
  ...
)

\method{proportion}{ps}(
  fitted_model,
  numerator_combination = NULL,
  denominator_combination = NULL,
  ...
)

\method{proportion}{rw}(
  fitted_model,
  numerator_combination = NULL,
  denominator_combination = NULL,
  ...
)

\method{proportion}{ps_single}(fitted_model, ...)

\method{proportion}{rw_single}(fitted_model, ...)
}
\arguments{
\item{fitted_model}{Fitted model object with appropriate class}

\item{numerator_combination}{Named pathogens or subtypes to be included in proportion numerator}

\item{denominator_combination}{Named pathogens or sutypes to be included in proportion denominator}

\item{...}{Additional arguments passed to methods}
}
\value{
Data frame with proportion analysis results
}
\description{
S3 generic for computing proportions from fitted models
}
