% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_influenza_stan.R
\name{ps_influenza_stan}
\alias{ps_influenza_stan}
\title{PS Influenza}
\usage{
ps_influenza_stan(
  num_data,
  num_knots,
  num_path,
  knots,
  spline_degree,
  Y,
  P1,
  P2,
  X,
  week_effect,
  DOW,
  cov_structure = c(0, 1, 2),
  ...
)
}
\arguments{
\item{num_data}{number of data points}

\item{num_knots}{num of knots}

\item{knots}{the sequence of knots}

\item{spline_degree}{the degree of spline (is equal to order - 1)}

\item{P1}{daily number of lab tests positive for influenza A (1st entry) and all other pathogens}

\item{P2}{daily number of influenza A H3N2, and influenza A H1N1}

\item{week_effect}{Number of days in day of week effect? 1=none, 2=weekends?, 7=all days}

\item{DOW}{integer of day of the week}

\item{cov_structure}{0 is tau\link{1}, 1 is tau\link{num_path}, 2 is Sigma\link{num_path, num_path}}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
PS Influenza
}
