% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_mp_stan.R
\name{ps_mp_stan}
\alias{ps_mp_stan}
\title{PS MP}
\usage{
ps_mp_stan(
  num_data,
  num_knots,
  num_path,
  knots,
  spline_degree,
  Y,
  P,
  X,
  week_effect,
  DOW,
  cov_structure = c(0, 1, 2),
  noise_structure = c(0, 1),
  ...
)
}
\arguments{
\item{num_data}{number of data points}

\item{num_knots}{number of knots}

\item{num_path}{number of pathogens}

\item{knots}{the sequence of knots}

\item{spline_degree}{the degree of spline (is equal to order - 1)}

\item{Y}{outcome variable, eg daily number of cases}

\item{P}{daily number of lab tests positive for each pathogen}

\item{X}{time data}

\item{week_effect}{Number of distinct days in day of week effect. 1 = single
effect shared by all days (essentially no DOW), 7 = each day a unique effect}

\item{DOW}{integer of day of the week}

\item{cov_structure}{0 is tau of 1, 1 is tau of
num_path, 2 is Sigma with two dimensions: num_path, num_path.}

\item{noise_structure}{0 only includes observation noise (same between
pathogens), 1 includes noise in individual pathogens as well}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
PS MP
}
