% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_model.R
\name{random_walk}
\alias{random_walk}
\title{Random walk method}
\usage{
random_walk()
}
\value{
list with method identified as random walk
}
\description{
Random walk method
}
\examples{
random_walk()
}
