% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rt.R
\name{Rt}
\alias{Rt}
\alias{Rt.ps}
\alias{Rt.rw}
\alias{Rt.ps_single}
\alias{Rt.rw_single}
\title{Generic Method for Rt Analysis}
\usage{
Rt(fitted_model, tau_max = 7, gi_dist, ...)

\method{Rt}{ps}(fitted_model, tau_max = 7, gi_dist, ...)

\method{Rt}{rw}(fitted_model, tau_max = 7, gi_dist, ...)

\method{Rt}{ps_single}(fitted_model, tau_max = 7, gi_dist, ...)

\method{Rt}{rw_single}(fitted_model, tau_max = 7, gi_dist, ...)
}
\arguments{
\item{fitted_model}{Fitted model object with appropriate class}

\item{tau_max}{Integer maximum generation interval in days (default: 7)}

\item{gi_dist}{Function that returns generation interval probability for given day}

\item{...}{Additional arguments passed to methods}
}
\value{
Data frame with Rt analysis results
}
\description{
S3 generic for computing reproduction numbers from fitted models
}
