% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rt.R
\name{Rt}
\alias{Rt}
\alias{Rt.ps}
\alias{Rt.rw}
\alias{Rt.ps_single}
\alias{Rt.rw_single}
\title{Generic Method for Rt Analysis}
\usage{
Rt(fitted_model, tau_max = 7, gi_dist, ...)

\method{Rt}{ps}(fitted_model, tau_max = 7, gi_dist, ...)

\method{Rt}{rw}(fitted_model, tau_max = 7, gi_dist, ...)

\method{Rt}{ps_single}(fitted_model, tau_max = 7, gi_dist, ...)

\method{Rt}{rw_single}(fitted_model, tau_max = 7, gi_dist, ...)
}
\arguments{
\item{fitted_model}{Fitted model object with class `EpiStrainDynamics.fit`}

\item{tau_max}{Integer maximum generation interval in days (default: 7)}

\item{gi_dist}{Function that returns generation interval probability for given day}

\item{...}{Additional arguments passed to metrics calculation}
}
\value{
named list of class `EpiStrainDynamics.metric` containing a dataframe
 of the calculated metric outcome (`$measure`), the fit object (`$fit`), and the
 constructed model object (`$constructed_model`). The `measure` data frame
 contains the median of the epidemiological quantity (`y`), the 50% credible
 interval of the quantity (`lb_50` & `ub_50`), the 95% credible interval
 (`lb_95` & `ub_95`), the proportion greater than a defined threshold value
 (`prop`), the pathogen name (`pathogen`), and the time label (`time`).
}
\description{
S3 generic for computing reproduction numbers from fitted models
}
\examples{
\dontrun{
  mod <- construct_model(
    method = random_walk(),
    pathogen_structure = single(
      case_timeseries = sarscov2$cases,
      time = sarscov2$date))

  fit <- fit_model(mod)

  rt <- Rt(fit, tau_max = 7, gi_dist = function(x) 4*x*exp(-2*x))
}

}
\seealso{
Other metrics: 
\code{\link{growth_rate}()},
\code{\link{incidence}()},
\code{\link{proportion}()}
}
\concept{metrics}
