% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw_single_stan.R
\name{rw_single_stan}
\alias{rw_single_stan}
\title{RW single}
\usage{
rw_single_stan(num_data, Y, week_effect, DOW, ...)
}
\arguments{
\item{num_data}{number of data points}

\item{Y}{daily number of 'cases'}

\item{week_effect}{number of days in day of week effect? 1=none, 2=weekends?, 7=all days}

\item{DOW}{integer of day of the week}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
RW single
}
