% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogen_structure.R
\name{single}
\alias{single}
\title{Single pathogen structure}
\usage{
single(case_timeseries, time, pathogen_name = "default")
}
\arguments{
\item{case_timeseries}{vector of total case data}

\item{time}{vector of labelled time data}

\item{pathogen_name}{optionally provided name of pathogen. This value is
preserved exactly as provided (case-sensitive). Default name is `default`.}
}
\value{
formatted list with pathogen structure and data of class
  `EpiStrainDynamics.pathogen_structure`.
}
\description{
Single pathogen structure
}
\examples{
single(
  case_timeseries = sarscov2$cases,
  time = sarscov2$date,
  pathogen_name = 'SARS-COV-2' # preserved as provided
)

single(
  case_timeseries = sarscov2$cases,
  time = sarscov2$date,
  pathogen_name = 'sarscov2' # preserved as provided
)

single(
  case_timeseries = sarscov2$cases,
  time = sarscov2$date
)
}
