% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogen_structure.R
\name{single}
\alias{single}
\title{Single pathogen structure}
\usage{
single(case_timeseries, time, pathogen_name = "default")
}
\arguments{
\item{case_timeseries}{vector of total case data}

\item{time}{vector of labelled time data}

\item{pathogen_name}{name of pathogen}
}
\value{
formatted list with pathogen structure and data
}
\description{
Single pathogen structure
}
\examples{
single(
  case_timeseries = sarscov2$cases,
  time = sarscov2$date,
  pathogen_name = 'SARS-COV-2'
)

}
