% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_model.R
\name{single}
\alias{single}
\title{Single pathogen structure}
\usage{
single(case_timeseries, pathogen_name)
}
\arguments{
\item{case_timeseries}{total case data}

\item{pathogen_name}{name of pathogen}
}
\value{
formatted list with pathogen structure and data
}
\description{
Single pathogen structure
}
