% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogen_structure.R
\name{single}
\alias{single}
\title{Single pathogen structure}
\usage{
single(data, case_timeseries, time = NULL)
}
\arguments{
\item{data}{dataframe containing columns with all relevant data, or a time
series object (ts, xts, zoo, tsibble, etc.)}

\item{case_timeseries}{Column name containing case counts. Must be numeric
or a \code{units} object from the \pkg{units} package.}

\item{time}{name of column with time data. Required for non-time-series
input data. Flexible format - can be date, index, or others, accepted
as \code{index} identifiers in the \code{tsibble} time format. Optional when
\code{data} is a time series class object (ts, mts, xts, zoo, zooreg, tsibble)
as the time index will be automatically detected.}
}
\value{
formatted list with pathogen structure and data of class
\code{EpiStrainDynamics.pathogen_structure}.
}
\description{
Single pathogen structure
}
\examples{
# Using a data frame
single(
  data = sarscov2,
  case_timeseries = 'cases',
  time = 'date'
)

# Using a time series object (time argument is optional)
\dontrun{
sarscov2_xts <- xts::xts(sarscov2[, c("cases", "alpha")], order.by = sarscov2$date)
single(
  data = sarscov2_xts,
  case_timeseries = 'cases'
)
}

}
\seealso{
Other pathogen_structure: 
\code{\link{multiple}()},
\code{\link{subtyped}()}
}
\concept{pathogen_structure}
