% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogen_structure.R
\name{subtyped}
\alias{subtyped}
\title{Subtyped pathogen structure}
\usage{
subtyped(
  case_timeseries,
  time,
  influenzaA_unsubtyped_timeseries,
  influenzaA_subtyped_timeseries,
  other_pathogen_timeseries,
  smoothing_structure = c("shared", "independent", "correlated"),
  observation_noise = c("observation_noise_only", "pathogen_specific_noise")
)
}
\arguments{
\item{case_timeseries}{vector of total case data}

\item{time}{vector of labelled time data}

\item{influenzaA_unsubtyped_timeseries}{case timeseries data for unsubtyped
influenzaA}

\item{influenzaA_subtyped_timeseries}{case timeseries data for subtyped
influenzaA}

\item{other_pathogen_timeseries}{other pathogen case timeseries}

\item{smoothing_structure}{either \code{shared} (all pathogens have the same
smoothing structure; \code{tau[1]}), \code{independent} (each pathogen with
independent smoothing structure; \code{tau[number of pathogens]}), or
\code{correlated} (smoothing structure is correlated among pathogens
\code{Sigma[number of pathogens, number of pathogens]}). Case-insensitive.}

\item{observation_noise}{either \code{observation_noise_only} (only includes
observation noise - the same between pathogens) or \code{pathogen_specific_noise}
(includes noise in individual pathogens as well). Case-insensitive.}
}
\value{
named list including pathogen_structure, pathogen_names, data,
and model_params of class \code{EpiStrainDynamics.pathogen_structure}
}
\description{
Subtyped pathogen structure
}
\examples{
# Both integer and numeric inputs work consistently
subtyped(
  case_timeseries = influenza$ili,
  time = influenza$week,
  influenzaA_unsubtyped_timeseries = influenza$inf_A,
  influenzaA_subtyped_timeseries = list(
    influenzaA.H3N2 = influenza$inf_H3N2,
    influenzaA.H1N1 = influenza$inf_H1N1
  ),
  other_pathogen_timeseries = list(
    influenzaB = influenza$inf_B,
    other = influenza$num_spec - influenza$inf_all
  ),
  smoothing_structure = 'independent',
  observation_noise = 'observation_noise_only'
)

}
\seealso{
Other pathogen_structure: 
\code{\link{multiple}()},
\code{\link{single}()}
}
\concept{pathogen_structure}
