% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_model.R
\name{subtyped}
\alias{subtyped}
\title{Subtyped pathogen structure}
\usage{
subtyped(
  case_timeseries,
  influenzaA_unsubtyped_timeseries,
  influenzaA_subtyped_timeseries,
  other_pathogen_timeseries,
  smoothing_structure = c("shared", "independent", "correlated"),
  observation_noise = c("observation_noise_only", "pathogen_specific_noise")
)
}
\arguments{
\item{case_timeseries}{total case data}

\item{influenzaA_unsubtyped_timeseries}{case timeseries data for unsubtyped
influenzaA}

\item{influenzaA_subtyped_timeseries}{case timeseries data for subtyped
influenzaA}

\item{other_pathogen_timeseries}{other pathogen case timeseries}

\item{smoothing_structure}{either 'shared' (all pathogens have the same;
tau\link{1}), 'independent' (each pathogen has completely independent smoothing
structure; tau\link{number of pathogens}), or 'correlated' (smoothing structure is
correlated among pathogens Sigma\link{number of pathogens, number of pathogens})}

\item{observation_noise}{either 'observation_noise_only' (only includes
observation noise - the same between pathogens) or pathogen_specific_noise'
(includes noise in individual pathogens as well)}
}
\value{
named list including pathogen_structure, data, and model_params
}
\description{
Subtyped pathogen structure
}
