% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogen_structure.R
\name{subtyped}
\alias{subtyped}
\title{Subtyped pathogen structure}
\usage{
subtyped(
  data,
  case_timeseries,
  time,
  influenzaA_unsubtyped_timeseries,
  influenzaA_subtyped_timeseries,
  other_pathogen_timeseries
)
}
\arguments{
\item{data}{dataframe containing columns with all relevant data}

\item{case_timeseries}{name of column with total case data, this name will be
used as the name of the pathogen}

\item{time}{name of column with time data. flexible format - can be date,
index, or others, accepted as \code{index} identifiers in the \code{tsibble} time
format.}

\item{influenzaA_unsubtyped_timeseries}{vector of column names with additional
unsubtyped influenzaA case count timeseries}

\item{influenzaA_subtyped_timeseries}{vector of column names with additional
subtyped influenzaA case count timeseries}

\item{other_pathogen_timeseries}{vector of column names with additional
pathogen case count timeseries to model}
}
\value{
named list including pathogen_structure, pathogen_names, and data
of class \code{EpiStrainDynamics.pathogen_structure}
}
\description{
Subtyped pathogen structure
}
\examples{
subtyped(
  data = influenza,
  case_timeseries = 'ili',
  time = 'week',
  influenzaA_unsubtyped_timeseries = 'inf_A',
  influenzaA_subtyped_timeseries = c('inf_H3N2', 'inf_H1N1'),
  other_pathogen_timeseries = c('inf_B', 'other')
)

}
\seealso{
Other pathogen_structure: 
\code{\link{multiple}()},
\code{\link{single}()}
}
\concept{pathogen_structure}
