% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogen_structure.R
\name{subtyped}
\alias{subtyped}
\title{Subtyped pathogen structure}
\usage{
subtyped(
  case_timeseries,
  time,
  influenzaA_unsubtyped_timeseries,
  influenzaA_subtyped_timeseries,
  other_pathogen_timeseries
)
}
\arguments{
\item{case_timeseries}{vector of total case data}

\item{time}{vector of labelled time data}

\item{influenzaA_unsubtyped_timeseries}{case timeseries data for unsubtyped
influenzaA}

\item{influenzaA_subtyped_timeseries}{case timeseries data for subtyped
influenzaA}

\item{other_pathogen_timeseries}{other pathogen case timeseries}
}
\value{
named list including pathogen_structure, pathogen_names, and data
of class \code{EpiStrainDynamics.pathogen_structure}
}
\description{
Subtyped pathogen structure
}
\examples{
# Both integer and numeric inputs work consistently
subtyped(
  case_timeseries = influenza$ili,
  time = influenza$week,
  influenzaA_unsubtyped_timeseries = influenza$inf_A,
  influenzaA_subtyped_timeseries = list(
    influenzaA.H3N2 = influenza$inf_H3N2,
    influenzaA.H1N1 = influenza$inf_H1N1
  ),
  other_pathogen_timeseries = list(
    influenzaB = influenza$inf_B,
    other = influenza$num_spec - influenza$inf_all
  )
)

}
\seealso{
Other pathogen_structure: 
\code{\link{multiple}()},
\code{\link{single}()}
}
\concept{pathogen_structure}
