% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-utils.R
\name{validate_class_inherits}
\alias{validate_class_inherits}
\title{Validate class of argument is correct}
\usage{
validate_class_inherits(obj, class_name)
}
\arguments{
\item{obj}{argument to check}

\item{class_name}{class name the object should be}
}
\description{
Validate class of argument is correct
}
